服务端能力完善，适配客户端接口

+++ src/main/java/com/veezean/codereview/server/service/ColumnDefineService.java
@@ -1,10 +1,13 @@
 package com.veezean.codereview.server.service;
 
 import cn.hutool.core.bean.BeanUtil;
+import cn.hutool.json.JSONUtil;
+import com.alibaba.fastjson.JSON;
 import com.veezean.codereview.server.common.CodeReviewException;
 import com.veezean.codereview.server.common.InputTypeDefine;
 import com.veezean.codereview.server.entity.ColumnDefineEntity;
-import com.veezean.codereview.server.model.SaveColumnDefineReqBody;
+import com.veezean.codereview.server.model.RecordColumns;
+import com.veezean.codereview.server.model.ValuePair;
 import com.veezean.codereview.server.repository.ColumnDefineRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
@@ -12,8 +15,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 /**
  * <类功能简要描述>
@@ -26,8 +32,10 @@ import java.util.stream.Collectors;
 public class ColumnDefineService {
     @Autowired
     private ColumnDefineRepository columnDefineRepository;
+    @Autowired
+    private DictService dictService;
 
-    public void createOrModifyColumn(long columnId, SaveColumnDefineReqBody reqBody) {
+    public void createOrModifyColumn(long columnId, ColumnDefineEntity reqBody) {
         reqValidate(reqBody);
         ColumnDefineEntity existColumn =
                 columnDefineRepository.findById(columnId).orElse(new ColumnDefineEntity());
@@ -35,7 +43,7 @@ public class ColumnDefineService {
         columnDefineRepository.saveAndFlush(existColumn);
     }
 
-    private void reqValidate(SaveColumnDefineReqBody reqBody) {
+    private void reqValidate(ColumnDefineEntity reqBody) {
         if (reqBody == null
                 || StringUtils.isEmpty(reqBody.getColumnCode())
                 || StringUtils.isEmpty(reqBody.getShowName())
@@ -58,22 +66,57 @@ public class ColumnDefineService {
 
     @Transactional
     public void deleteColumn(long columnId) {
-        columnDefineRepository.deleteById(columnId);
+        ColumnDefineEntity columnDefineEntity = columnDefineRepository.findById(columnId).orElse(null);
+        if (columnDefineEntity != null) {
+            if (columnDefineEntity.isSystemInitialization()) {
+                throw new CodeReviewException("系统预置字段，不允许删除");
+            }
+            columnDefineRepository.deleteById(columnId);
+        }
     }
 
     @Transactional
     public void deleteColumns(List<Long> columnIds) {
-        for (Long columnId : columnIds) {
-            columnDefineRepository.deleteById(columnId);
+        boolean containsSystemField = columnDefineRepository.findAllById(columnIds).stream()
+                .anyMatch(ColumnDefineEntity::isSystemInitialization);
+        if (containsSystemField) {
+            throw new CodeReviewException("包含系统预置字段，不允许删除");
         }
+        columnDefineRepository.deleteAllById(columnIds);
     }
 
     public ColumnDefineEntity queryColumn(long columnId) {
         return columnDefineRepository.findById(columnId).orElseThrow(() -> new CodeReviewException("数据不存在"));
     }
 
-    public List<ColumnDefineEntity> queryColumns() {
-        return columnDefineRepository.findAll().stream().sorted((o1, o2) -> o1.getSortIndex() - o2.getSortIndex() > 0 ? 1 : -1)
-                .collect(Collectors.toList());
+    public ColumnDefineEntity queryColumnByCode(String code) {
+        return columnDefineRepository.findFirstByColumnCode(code);
+    }
+
+    public Stream<ColumnDefineEntity> queryColumns() {
+        return columnDefineRepository.findAll().stream()
+                .sorted(Comparator.comparingInt(ColumnDefineEntity::getSortIndex))
+                .peek(columnDefineEntity -> {
+                    if (InputTypeDefine.COMBO_BOX.getValue().equals(columnDefineEntity.getInputType())
+                            && StringUtils.isNotEmpty(columnDefineEntity.getDictCollectionCode())) {
+                        List<ValuePair> items =
+                                dictService.queryListItemsByCollectionCode(columnDefineEntity.getDictCollectionCode());
+                        columnDefineEntity.setEnumValues(items);
+                    }
+                });
+    }
+
+    public List<ValuePair> queryFieldDictItems(String fieldCode) {
+        ColumnDefineEntity entity = columnDefineRepository.findFirstByColumnCode(fieldCode);
+        if (entity != null && StringUtils.isNotEmpty(entity.getDictCollectionCode())) {
+            return dictService.queryListItemsByCollectionCode(entity.getDictCollectionCode());
+        }
+        return new ArrayList<>();
+    }
+
+    public String getJsonContent() {
+        RecordColumns columns = new RecordColumns();
+        columns.setColumns(queryColumns().collect(Collectors.toList()));
+        return JSONUtil.formatJsonStr(JSON.toJSONString(columns));
     }
 }
